requirements.txt
scipy==1.6.2
pandas==1.2.4
plotly==4.14.3
numpy==1.20.2
redditcleaner==1.1.2
dash==1.20.0
dash-bootstrap-components==0.12.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Pillow==8.2.0
gunicorn==20.1.0
matplotlib~=3.4.1
nltk~=3.6.2
wordcloud~=1.8.1
requests~=2.25.1
pytz~=2021.1
urllib3~=1.26.4